refactor : pr리뷰 수정 (#49)
